Refactor data loader
