+++ extract_nasnet_features.py
@@ -2,7 +2,7 @@ import tensorflow as tf
 from keras import backend as K
 
 from utils.nasnet import NASNetMobile
-from data_loader import train_generator, val_generator
+from utils.data_loader import train_generator, val_generator
 
 sess = tf.Session()
 K.set_session(sess)
